refactor folder/file organization
