+++ tests/mcp_tests/test_semantic_tool_filter_e2e.py
@@ -18,7 +18,7 @@ from mcp.types import Tool as MCPTool
 async def test_e2e_semantic_filter():
     """E2E: Load router/filter and verify hook filters tools."""
     from litellm import Router
-    from litellm.proxy.hooks.semantic_tool_filter_hook import SemanticToolFilterHook
+    from litellm.proxy.hooks.mcp_semantic_filter import SemanticToolFilterHook
     from litellm.proxy._experimental.mcp_server.semantic_tool_filter import (
         SemanticMCPToolFilter,
     )
